requirements.txt
streamlit>=1.36.0
yt-dlp>=2024.10.7
requests>=2.32.0
openai>=1.40.0
notion-client>=2.2.1
python-dotenv>=1.0.1
